+++ docs/history.md
@@ -7,6 +7,7 @@
 
 - Changed the following babies:
   - Lowface Baby (#73) - Reveals the floor layout after 6 hits
+  - Twotone Baby (#346) - Spending bombs uses up keys first
 - Fixed the bug where Tabby Baby was not working properly. (Thanks Fallenfor)
 - Fixed the bug where Blue Baby was not working properly.
 - Fixed the bug where Empty Vessel would result in a bugged baby costume. (Thanks Gamonymous)